feat : 클래스 네이밍 수정, 유효성 검사 후에 초기화 하도록 순서 수정, checked exception 제거
